requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-genai==0.5.0